improve test cleanup

+++ tests/_tools.js
@@ -20,7 +20,7 @@ export const fileCleanup = (testOutputDirName, fileNamePartial) => {
             if (err) throw err
 
             for (const file of files) {
-                if (file.includes(fileNamePartial)) {
+                if (file.includes(fileNamePartial) &&  fs.existsSync(path.join(__dirname, '..', file))) {
                     fs.unlink(path.join(__dirname, '..', file), err => {
                         if (err) throw err
                     })
